focused_test.py
#!/usr/bin/env python3
"""
Focused test for enhanced features that don't depend on Gemini API
"""

import requests
import json
import tempfile
import os
from reportlab.pdfgen import canvas
from reportlab.lib.pagesizes import letter
from docx import Document

BACKEND_URL = "https://f800b529-c05c-42da-a95a-92830f186bd2.preview.emergentagent.com/api"

def test_enhanced_health_check():
    """Test enhanced health check endpoint"""
    print("Testing Enhanced Health Check...")
    response = requests.get(f"{BACKEND_URL}/health", timeout=10)
    
    if response.status_code == 200:
        data = response.json()
        features = data.get("features", {})
        
        # Check all enhanced features
        auth_enabled = features.get("authentication") == True
        doc_parsing = set(features.get("document_parsing", [])) >= {"PDF", "DOCX", "TXT", "MD"}
        analysis_types = set(features.get("analysis_types", [])) >= {"gap_analysis", "requirements_extraction", "summary"}
        
        if auth_enabled and doc_parsing and analysis_types:
            print("✅ Enhanced health check passed - all features configured")
            return True
        else:
            print(f"❌ Missing features: auth={auth_enabled}, docs={doc_parsing}, analysis={analysis_types}")
            return False
    else:
        print(f"❌ Health check failed: {response.status_code}")
        return False

def test_authentication_system():
    """Test complete authentication system"""
    print("Testing Authentication System...")
    
    # Test user registration/login
    user_data = {
        "email": "test.user@example.com",
        "password": "TestPass123!",
        "name": "Test User"
    }
    
    session = requests.Session()
    
    # Try registration (might fail if user exists)
    reg_response = session.post(f"{BACKEND_URL}/auth/register", json=user_data, timeout=10)
    
    if reg_response.status_code == 400:  # User exists, try login
        login_response = session.post(f"{BACKEND_URL}/auth/login", 
                                    json={"email": user_data["email"], "password": user_data["password"]}, 
                                    timeout=10)
        if login_response.status_code != 200:
            print(f"❌ Login failed: {login_response.status_code}")
            return False
        token_data = login_response.json()
    elif reg_response.status_code == 200:
        token_data = reg_response.json()
    else:
        print(f"❌ Registration failed: {reg_response.status_code}")
        return False
    
    # Set auth header
    token = token_data["access_token"]
    session.headers.update({"Authorization": f"Bearer {token}"})
    
    # Test protected route
    me_response = session.get(f"{BACKEND_URL}/auth/me", timeout=10)
    if me_response.status_code != 200:
        print(f"❌ Protected route failed: {me_response.status_code}")
        return False
    
    # Test protected route without auth
    temp_session = requests.Session()
    unauth_response = temp_session.get(f"{BACKEND_URL}/auth/me", timeout=10)
    if unauth_response.status_code not in [401, 403]:
        print(f"❌ Unprotected route issue: {unauth_response.status_code}")
        return False
    
    print("✅ Authentication system working correctly")
    return True, session

def test_document_parsing_capabilities():
    """Test document parsing without Gemini API dependency"""
    print("Testing Document Parsing Capabilities...")
    
    # Test unsupported file type
    temp_file = tempfile.NamedTemporaryFile(mode='w', suffix='.xyz', delete=False)
    temp_file.write("This is an unsupported file type")
    temp_file.close()
    
    try:
        with open(temp_file.name, 'rb') as file:
            files = {'file': ('test.xyz', file, 'application/octet-stream')}
            data = {'analysis_type': 'gap_analysis'}
            
            response = requests.post(f"{BACKEND_URL}/analyze-document-file", files=files, data=data, timeout=10)
            
            if response.status_code == 400 and "Unsupported file type" in response.text:
                print("✅ Unsupported file type correctly rejected")
                return True
            else:
                print(f"❌ Unsupported file handling failed: {response.status_code}")
                return False
    finally:
        os.unlink(temp_file.name)

def test_analysis_history_features():
    """Test analysis history and user-specific features"""
    print("Testing Analysis History Features...")
    
    # Test public analyses access
    public_response = requests.get(f"{BACKEND_URL}/analyses", timeout=10)
    if public_response.status_code != 200:
        print(f"❌ Public analyses access failed: {public_response.status_code}")
        return False
    
    public_analyses = public_response.json()
    print(f"✅ Public analyses accessible: {len(public_analyses)} analyses")
    
    # Test with authentication (if available)
    try:
        auth_result = test_authentication_system()
        if isinstance(auth_result, tuple):
            _, session = auth_result
            
            # Test user-specific analyses
            user_response = session.get(f"{BACKEND_URL}/analyses", timeout=10)
            if user_response.status_code == 200:
                user_analyses = user_response.json()
                print(f"✅ User-specific analyses accessible: {len(user_analyses)} analyses")
                
                # Test specific analysis retrieval
                if user_analyses:
                    analysis_id = user_analyses[0]["id"]
                    specific_response = session.get(f"{BACKEND_URL}/analyses/{analysis_id}", timeout=10)
                    if specific_response.status_code == 200:
                        print("✅ Specific analysis retrieval working")
                        return True
                    else:
                        print(f"❌ Specific analysis retrieval failed: {specific_response.status_code}")
                        return False
                else:
                    print("✅ User-specific analysis system working (no analyses to test)")
                    return True
            else:
                print(f"❌ User-specific analyses failed: {user_response.status_code}")
                return False
    except:
        print("⚠️  Authentication not available for user-specific testing")
        return True

def main():
    """Run all focused tests"""
    print("=" * 60)
    print("FOCUSED ENHANCED FEATURES TEST")
    print("=" * 60)
    
    tests = [
        test_enhanced_health_check,
        test_authentication_system,
        test_document_parsing_capabilities,
        test_analysis_history_features
    ]
    
    results = []
    for test in tests:
        try:
            result = test()
            if isinstance(result, tuple):
                result = result[0]  # Extract boolean from tuple
            results.append(result)
        except Exception as e:
            print(f"❌ Test failed with exception: {e}")
            results.append(False)
        print()
    
    passed = sum(results)
    total = len(results)
    
    print("=" * 60)
    print(f"FOCUSED TEST SUMMARY: {passed}/{total} tests passed")
    print("=" * 60)
    
    return passed == total

if __name__ == "__main__":
    success = main()
    exit(0 if success else 1)